// src/question-group/enums/question-type.enum.ts
export enum QuestionType {
  IDENTIFYING_INFORMATION = 'IDENTIFYING_INFORMATION',
  IDENTIFYING_VIEWS_CLAIMS = 'IDENTIFYING_VIEWS_CLAIMS',
  NOTE_COMPLETION = 'NOTE_COMPLETION',
  MATCHING_HEADINGS = 'MATCHING_HEADINGS',
  SUMMARY_COMPLETION = 'SUMMARY_COMPLETION',
  MULTIPLE_CHOICE_1_ANSWER = 'MULTIPLE_CHOICE_1_ANSWER',
  MULTIPLE_CHOICE_MULTIPLE_ANSWER = 'MULTIPLE_CHOICE_MULTIPLE_ANSWER',
  MATCHING_SENTENCE_ENDINGS = 'MATCHING_SENTENCE_ENDINGS',
  LINE_GRAPH = 'LINE_GRAPH',
  CAUSES_SOLUTIONS = 'CAUSES_SOLUTIONS',
  SENTENCE_COMPLETION = 'SENTENCE_COMPLETION',
  LABELLING_A_PLAN_MAP = 'LABELLING_A_PLAN_MAP',
  FLOW_CHART_COMPLETION = 'FLOW_CHART_COMPLETION',
  FORM_COMPLETION = 'FORM_COMPLETION',
  TABLE_COMPLETION = 'TABLE_COMPLETION',
  SHORT_ANSWER_QUESTION = 'SHORT_ANSWER_QUESTION',
  DIAGRAM_LABELING = 'DIAGRAM_LABELING',
  PICK_FROM_A_LIST = 'PICK_FROM_A_LIST',
  MATCHING_PARAGRAPH_INFORMATION = 'MATCHING_PARAGRAPH_INFORMATION',
  MATCHING_FEATURES = 'MATCHING_FEATURES',
  LIST_SELECTION = 'LIST_SELECTION',
}